requirements.txt
astroid==2.3.0
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.3
pylint==2.4.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
